add support for wasm
